rmp packages preun script launches only if last package removed
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@884 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/package/afanasy-common/RPM/SPECS/afanasy-common.spec
@@ -26,8 +26,15 @@ done
 
 %files
 %defattr(-,root,root)
-/opt
+/opt/cgru/afanasy/bin
+/opt/cgru/afanasy/init
+/opt/cgru/afanasy/python
+/opt/cgru/afanasy/config_default.xml
+/opt/cgru/afanasy/farm_default.xml
+/opt/cgru/afanasy/pathmap.txt
+/opt/cgru/afanasy/setup.sh
 %config /opt/cgru/afanasy/config.xml
+%config /opt/cgru/afanasy/nonrootuser
 
 %clean
 
@@ -37,7 +44,9 @@ echo "Afanasy common POST INSTALL"
 exit 0
 
 %preun
-echo "Afanasy common PRE REMOVE:"
+echo "Afanasy common PRE REMOVE: $1"
+[ "$1" != "0" ] && exit 0
 pkill afcmd || true
+echo "Cleaning afanasy/python"
 [ -d /opt/cgru/afanasy/python ] && find /opt/cgru/afanasy/python -type f -name *.pyc -exec rm -vf {} \;
 exit 0
