+++ afanasy/trunk/package/afanasy-render/RPM/SPECS/afanasy-render.spec
@@ -48,7 +48,8 @@ id renderer || useradd renderer --create-home
 exit 0
 
 %preun
-echo "Afanasy render PRE REMOVE:"
+echo "Afanasy render PRE REMOVE: $1"
+[ "$1" != "0" ] && exit 0
 /etc/init.d/afrender stop || true
 /opt/cgru/afanasy/init/initlinks.sh c rm afrender || true
 exit 0
